PAL-999 refactored to remove some intermediary packages
