+++ hadoop-reader/src/main/java/uk/gov/gchq/palisade/service/data/reader/hadoop/HadoopDataReader.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.gchq.palisade.reader;
+package uk.gov.gchq.palisade.service.data.reader.hadoop;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
